requirements.txt
Django==2.2.3
django-coverage==1.2.4
django-pytest==0.2.0
packaging==19.0
tox==3.13.2
flake8==3.7.8
psycopg2==2.8.3




